requirements.txt
simplejson~=3.17.2
discord.py~=1.4.1